# openapi.yaml
openapi: 3.0.3
info:
  title: SaveState Retro Backlog API
  version: 1.0.0
  description: API for a retro game catalog, personal backlog, and reviews.
servers:
  - url: /api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string, nullable: true }
              message: { type: string }
      required: [success, message]
    SuccessEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: {}
      required: [success, data]
...
    LoginRequest:
      type: object
      properties:
        email: { type: string, format: email, example: admin@savestate.dev }
        password: { type: string, example: admin123 }
      required: [email, password]
    LoginResponseData:
      type: object
      properties:
        token: { type: string }
        user:
          type: object
          properties:
            id: { type: integer, example: 1 }
            email: { type: string, format: email }
            display_name: { type: string, example: Admin }
            role: { type: string, example: admin }
          required: [id, email, display_name, role]
      required: [token, user]
    User:
      type: object
      properties:
        id: { type: integer, example: 1 }
        email: { type: string, format: email }
        display_name: { type: string, example: Admin }
        role: { type: string, example: admin }
        is_active: { type: boolean, example: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time, nullable: true }
      required: [id, email, display_name, role, is_active, created_at]
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login (devuelve JWT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginResponseData"
        "400":
          description: Validación
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/me:
    get:
      tags: [Auth]
      summary: Usuario actual
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user: { $ref: "#/components/schemas/User" }
                        required: [user]
        "401":
          description: No autenticado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...